Shift method readFile from Class Task to Class Storage
