+++ data/output.txt
@@ -1,4 +1,3 @@
-T | 1 | shit
-T | 0 | go delta
+T | 1 | go delta
 E | 0 | coaching | 6pm | 7pm
-T | 1 | leave
+T | 1 | grab hockey stick
