+++ src/main/java/omoh/Storage.java
@@ -6,6 +6,7 @@ import omoh.tasktypes.Task;
 
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.util.Scanner;
 
 public class Storage {
 
@@ -24,7 +25,7 @@ public class Storage {
         //code to create new file in data directory
         File f = new File("data/output.txt");
         try {
-            Task.readFile();
+            readFile();
         } catch (FileNotFoundException e) {
             System.out.println("File not found");
         } catch (EmptyTodoException e) {
@@ -32,5 +33,27 @@ public class Storage {
         } catch (CorruptedFileException e) {
             System.out.println("Output.txt file format is wrong. File corrupted.");
         }
+    }
+
+
+    //Method to read and process the output.txt file
+    public static void readFile() throws FileNotFoundException, CorruptedFileException, EmptyTodoException {
+        //open file for reading
+        File f = new File("data/output.txt");
+        Scanner s = new Scanner(f);
+        Task.initArray();
+        //iteration required so that programme knows which task to mark as done
+        int iteration = 0;
+        while (s.hasNextLine()) {
+            String line = s.nextLine();
+            // Process each line (splitting by "|", for example)
+            String[] parts = line.split("\\|"); // Split the line by "|"
+            if (parts.length < 3 || parts.length > 5 ) {
+                throw new CorruptedFileException();
+            }
+            Task.processFileText(parts, iteration);
+            iteration++;
+        }
+        s.close();
     }
 }
